requirements/base.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.0
sqlparse==0.5.3
django-filter==25.1
boto3==1.39.12
botocore==1.39.12
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
urllib3==2.5.0
django-environ==0.12.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
